% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceylon_data_2.R
\docType{data}
\name{district}
\alias{district}
\title{District level shape file data and estimated estimated population (2020)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 4 columns.
}
\source{
https://data.humdata.org/dataset/sri-lanka-administrative-levels-0-4-boundaries

https://www.citypopulation.de/en/srilanka/prov/admin/
}
\usage{
district
}
\description{
shape file data and population data for district level
\describe{
\item{DISTRICT}{district name}
\item{geometry}{multipolygon}
\item{population}{district population}}
}
\keyword{datasets}
