% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceylon_data_1.R
\docType{data}
\name{province}
\alias{province}
\title{Province level shape file data and estimated population statistic}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.
}
\source{
https://data.humdata.org/dataset/sri-lanka-administrative-levels-0-4-boundaries
}
\usage{
province
}
\description{
shape file data for province level and estimated population statistic
\describe{
\item{PROVINCE}{name of the province}
\item{geometry}{Multipolygon}
\item{population}{Province level estimated population}}
}
\keyword{datasets}
