% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivers.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Rivers and Streams in Sri Lanka shape file data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 6969 rows and 3 columns.
}
\source{
https://data.humdata.org/dataset/sri-lanka-water-bodies-0-0?#
}
\usage{
rivers
}
\description{
shape file data for rivers and streams
\describe{
\item{CODE}{River code}
\item{NAME}{River name}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
