% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceylon_data_0.R
\docType{data}
\name{sf_sl_0}
\alias{sf_sl_0}
\title{Province level shape file data}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 2 columns.
}
\source{
https://data.humdata.org/dataset/sri-lanka-administrative-levels-0-4-boundaries
}
\usage{
sf_sl_0
}
\description{
shape file data for province level
\describe{
\item{COUNTRY}{country name is Sri Lanka}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
