% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceylon_data_2.R
\docType{data}
\name{sf_sl_2}
\alias{sf_sl_2}
\title{Province level shape file data}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 2 columns.
}
\source{
https://data.humdata.org/dataset/sri-lanka-administrative-levels-0-4-boundaries
}
\usage{
sf_sl_2
}
\description{
shape file data for district level
\describe{
\item{DISTRICT}{district name}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
