% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceylon_data_3.R
\docType{data}
\name{sf_sl_3}
\alias{sf_sl_3}
\title{Province level shape file data}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 333 rows and 2 columns.
}
\source{
https://data.humdata.org/dataset/sri-lanka-administrative-levels-0-4-boundaries
}
\usage{
sf_sl_3
}
\description{
shape file data for province level
\describe{
\item{ADM3_EN}{Divisional secretariat}
\item{geometry}{Multipolygon}}
}
\keyword{datasets}
